backend/pyproject.toml
[project]
name = "portfolio-api"
version = "0.1.0"
description = "REST API for self-hosted developer portfolio — FastAPI + SQLModel"
requires-python = ">=3.12.8"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.128.7",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlmodel>=0.0.32",
    "structlog>=25.5.0",
]

[dependency-groups]
dev = [
    "bandit>=1.9.3",
    "mypy>=1.19.1",
    "pip-audit>=2.10.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.0",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort (import sorting)
    "N",    # pep8-naming
    "UP",   # pyupgrade (modern Python syntax)
    "B",    # flake8-bugbear (common bugs)
    "SIM",  # flake8-simplify
    "RUF",  # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (le formatter s'en charge)
]

[tool.ruff.lint.per-file-ignores]
"**/routers/*" = ["B008", "E712"]
"**/app/auth.py" =["B008"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
plugins = ["pydantic.mypy"]
python_version = "3.12"
cache_dir = "backend/.mypy_cache"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "--cov=app --cov-report=term-missing --verbose"

[tool.coverage.run]
source = ["app"]
omit = ["tests/*"]

[tool.coverage.report]
fail_under = 85
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.",
    "if TYPE_CHECKING:",
]

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101"]
